+++ app/src/views/RegView.vue
@@ -93,6 +93,7 @@ watch(
 </script>
 
 <template>
+  <!-- Show the header at the top of the page -->
   <Header class="h-11" @toggle-menu="showMenu = !showMenu" />
 
   <div class="flex h-full flex-row">
@@ -102,9 +103,12 @@ watch(
       class="w-[21rem] bg-white pb-1"
       :class="!showMenu ? 'hidden' : ''"
     />
+
+    <!-- Show the main body and fill the remaining screen space -->
     <div class="mt-4 flex-grow overflow-y-scroll">
       <div class="m-auto max-w-[1200px]">
         <div v-if="reg">
+          <!-- Show the field/register value encode/decode table and associated buttons -->
           <RegLayout
             v-if="reg.fields"
             :fields="reg.fields"
@@ -115,11 +119,13 @@ watch(
             @stop-highlight-field="stopHighlightField"
           />
 
+          <!-- Show the register doc description -->
           <div v-if="doc" class="m-auto mx-4 mt-4">
             <!-- eslint-disable-next-line vue/no-v-html -->
             <span class="default-styles" v-html="doc"></span>
           </div>
 
+          <!-- Show the register field description table -->
           <RegFields
             v-if="reg.fields"
             :fields="reg.fields"